Update hu-hm2177-fb046r.txt
Here are my changes/additions. I've represented the strikeouts as [-content-]. There may be a better way to code this, but I'm not sure. Thoughts? I've left ??? as placeholders for the words I couldn't figure out. I'll come back to this again later in the weekend to see if I can make better sense of it.